+++ packages/steiner-cli/templates/app/src/App.js
@@ -6,10 +6,11 @@ import LoadingBar from 'react-redux-loading-bar';
 import Sidebar from 'react-sidebar';
 import Helmet from 'react-helmet';
 import theme from 'reapop-theme-wybo';
-import { MatchWhenAuthorized, MatchWhenGuest, HeaderLink, auth, createConfirm } from 'steiner';
+import { ControlledRouter, MatchWhenAuthorized, MatchWhenGuest, HeaderLink, auth, createConfirm } from 'steiner';
 import { getUser } from 'steiner/lib/auth/reducer';
 import ReactCSSTransitionGroup from 'react-addons-css-transition-group';
 import Dropdown from 'vivi/lib/Dropdown';
+import { getRouter } from 'steiner/lib/routing/reducer';
 
 import routeRegister from 'helpers/routeRegister';
 import Welcome from './components/Welcome';
@@ -22,6 +23,7 @@ import Settings from './components/Settings';
 import Profile from './components/Profile';
 import ModalHelp from './components/ModalHelp';
 import routes from './routes';
+import history from './history';
 
 import KeyBinderHoc from './components/KeyBinder';
 
@@ -86,10 +88,15 @@ class App extends Component {
     }
 
     render() {
-        const { user } = this.props;
+        const { user, router, dispatch } = this.props;
 
         return (
-            <BrowserRouter>
+            <ControlledRouter
+                history={history}
+                location={router.location}
+                action={router.action}
+                dispatch={dispatch}
+            >
                 <div>
                     <Helmet
                         title="App"
@@ -159,11 +166,11 @@ class App extends Component {
                         }}
                     />
                 </div>
-            </BrowserRouter>
+            </ControlledRouter>
         );
     }
 }
 
 const KeyedApp = KeyBinderHoc(App);
 
-export default connect(state => ({ user: getUser(state) }))(KeyedApp);
+export default connect(state => ({ user: getUser(state), router: getRouter(state) }))(KeyedApp);