+++ packages/steiner-cli/templates/app/src/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import { reducer as formReducer } from 'redux-form';
 import { reducer as notificationsReducer } from 'reapop';
 import { loadingBarReducer } from 'react-redux-loading-bar';
-import { auth } from 'steiner';
+import { auth, routing } from 'steiner';
 
 // import posts from '../modules/posts/reducers/posts';
 
@@ -11,7 +11,8 @@ const rootReducer = combineReducers({
     loadingBar: loadingBarReducer,
     notifications: notificationsReducer,
     // posts,
-    user: auth.reducer
+    user: auth.reducer,
+    router: routing.reducer
 });
 
 export default rootReducer;