+++ packages/steiner/src/routing/reducer.js
@@ -1,18 +1,24 @@
+import Immutable from 'seamless-immutable';
+
 import { actionTypes } from './actions';
 import { createReducer } from '../helpers/reducerCreator';
 
-export const DEFAULT_STATE = {
-    location: null,
-    action: null
-};
+export const DEFAULT_STATE = Immutable({
+    location: { pathname: '/' },
+    action: 'PUSH'
+});
 
 const handlers = {
     [actionTypes.navigate]: (state, action) => {
-        return {
+        return Immutable({
             location: action.location,
             action: action.action
-        };
+        });
     }
 };
 
 export default createReducer(handlers, DEFAULT_STATE);
+
+export function getRouter(state) {
+    return state.router;
+}