+++ packages/steiner-cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.8.0 (October 20, 2016)
+
+- Update templates for steiner@0.8.0
+
 ## 0.7.0 (October 18, 2016)
 
 - Updated templates for `SteinerHelper`