+++ packages/steiner/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.0 (October 20, 2016)
 
-- Added routing helpers
+- Added <ControlledRouter /> component
+- Added sync between list filters and url (experimental)
 
 ## 0.7.0 (October 18, 2016)